Add MapKit

+++ app/src/main/java/com/example/afishaapp/data/module/Coordinate.kt
@@ -1,6 +1,6 @@
 package com.example.afishaapp.data.module
 
 data class Coordinate(
-    val lat: Float,
-    val lon: Float
+    val lat: Double,
+    val lon: Double
 )
